feat!: metadata with custom NTP timestamp
The metadatahandler is split into a `parser` and `cb`
function. The former can modify the message that is sent
to the scheduler, so that custom NTP timestamps can be
used instead of those that are determined by the RTP
timestamp.

The `cb` is a synchonized callback that is run whenever
the parsed message is in sync with the video. This is
the former `metadataHandler` function.

BREAKING CHANGE: the optional `metadataHandler` property
has changed to an object {parser, cb} instead of a simple
function. If you use this property, consult the TypeScript
documentation for the new interface.