feat: progress bar
Add progress indication to the controls section.
The bar indicates played and buffered time, and
has a label with either "LIVE" for infinite duration
or "played / total" for a finite duration.

Clicking on the progress bar sets a new starting
offset in seconds, which only takes effect when
using RTSP streams.